General-Refactoring - Got the project to work; Switched to multi module; Made changes to model to work with: JDBC, JSON, Message Broker; Added postman collection for testing and documentation
