+++ Bookings/bookings-core/src/main/java/org/ecabs/bookings/infrastructure/db/AuditRepository.java
@@ -1,0 +1,9 @@
+package org.ecabs.bookings.infrastructure.db;
+
+import org.ecabs.bookings.model.AuditEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AuditRepository  extends JpaRepository<AuditEntity, Long> {
+}
