+++ Bookings/bookings-core/src/main/java/org/ecabs/bookings/infrastructure/db/BookingRepository.java
@@ -1,0 +1,14 @@
+package org.ecabs.bookings.infrastructure.db;
+
+import org.ecabs.bookings.model.Booking;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.UUID;
+
+@Repository
+public interface BookingRepository extends JpaRepository<Booking, Long> {
+
+    List<Booking> findAllByBookingId(UUID bookingId);
+}
