+++ Bookings/bookings-core/src/main/java/org/ecabs/bookings/infrastructure/db/WaypointRepository.java
@@ -1,0 +1,9 @@
+package org.ecabs.bookings.infrastructure.db;
+
+import org.ecabs.bookings.model.TripWaypoint;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface WaypointRepository  extends JpaRepository<TripWaypoint, Long> {
+}
