+++ Bookings/bookings-model/src/main/java/org/ecabs/bookings/utils/CustomInstantDeserializer.java
@@ -1,0 +1,21 @@
+package org.ecabs.bookings.utils;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+import java.io.IOException;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+public class CustomInstantDeserializer extends JsonDeserializer<Instant> {
+
+    private DateTimeFormatter fmt = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").withZone(ZoneId.of("Europe/Malta"));
+
+    @Override
+    public Instant deserialize(JsonParser p, DeserializationContext ctxt) throws IOException, JsonProcessingException {
+        return Instant.from(fmt.parse(p.getText()));
+    }
+}
