+++ Bookings/bookings-model/src/main/java/org/ecabs/bookings/utils/CustomInstantSerializer.java
@@ -1,0 +1,23 @@
+package org.ecabs.bookings.utils;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+public class CustomInstantSerializer extends JsonSerializer<Instant> {
+
+    private DateTimeFormatter fmt = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").withZone(ZoneId.of("Europe/Malta"));
+
+    @Override
+    public void serialize(Instant value, JsonGenerator gen, SerializerProvider serializers) throws IOException, JsonProcessingException {
+        String str = fmt.format(value);
+
+        gen.writeString(str);
+    }
+}
