+++ Bookings/bookings-model/src/main/java/org/ecabs/bookings/utils/CustomOffsetDateTimeDeserializer.java
@@ -1,0 +1,23 @@
+package org.ecabs.bookings.utils;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+
+import java.io.IOException;
+import java.time.OffsetDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+
+//Spring was complaining about deserialization on OffsetDateTime
+public class CustomOffsetDateTimeDeserializer extends JsonDeserializer<OffsetDateTime> {
+
+    private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").withZone(ZoneId.of("Europe/Malta"));
+
+    @Override
+    public OffsetDateTime deserialize(JsonParser parser, DeserializationContext context) throws IOException, JsonProcessingException {
+        return ZonedDateTime.parse(parser.getText(), this.formatter).toOffsetDateTime();
+    }
+}
