+++ Bookings/bookings-model/src/main/java/org/ecabs/bookings/utils/CustomOffsetDateTimeSerializer.java
@@ -1,0 +1,21 @@
+package org.ecabs.bookings.utils;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+import java.time.OffsetDateTime;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+
+//Spring was complaining about serialization on OffsetDateTime
+public class CustomOffsetDateTimeSerializer extends JsonSerializer<OffsetDateTime> {
+    private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss").withZone(ZoneId.of("Europe/Malta"));
+
+    @Override
+    public void serialize(OffsetDateTime value, JsonGenerator gen, SerializerProvider provider) throws IOException, JsonProcessingException {
+        gen.writeString(value.format(this.formatter));
+    }
+}
