+++ Bookings/bookings-ws/src/main/java/Bookings.java
@@ -1,0 +1,14 @@
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EnableJpaRepositories
+public class Bookings {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Bookings.class, args);
+	}
+
+}
