--- Bookings/src/main/java/org/ecabs/bookings/Bookings.java
@@ -1,13 +1,0 @@
-package org.ecabs.bookings;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Bookings {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Bookings.class, args);
-	}
-
-}
