--- Bookings/src/main/java/org/ecabs/bookings/infrastructure/CustomInstantDeserializer.java
@@ -1,21 +1,0 @@
-package org.ecabs.bookings.infrastructure;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-
-import java.io.IOException;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-
-public class CustomInstantDeserializer extends JsonDeserializer<Instant> {
-
-    private DateTimeFormatter fmt = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSSSSS").withZone(ZoneId.of("UTC"));
-
-    @Override
-    public Instant deserialize(JsonParser p, DeserializationContext ctxt) throws IOException, JsonProcessingException {
-        return Instant.from(fmt.parse(p.getText()));
-    }
-}
