--- Bookings/src/main/java/org/ecabs/bookings/infrastructure/CustomInstantSerializer.java
@@ -1,23 +1,0 @@
-package org.ecabs.bookings.infrastructure;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-
-import java.io.IOException;
-import java.time.Instant;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-
-public class CustomInstantSerializer extends JsonSerializer<Instant> {
-
-    private DateTimeFormatter fmt = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSSSSS").withZone(ZoneId.of("UTC"));
-
-    @Override
-    public void serialize(Instant value, JsonGenerator gen, SerializerProvider serializers) throws IOException, JsonProcessingException {
-        String str = fmt.format(value);
-
-        gen.writeString(str);
-    }
-}
