--- Bookings/src/main/java/org/ecabs/bookings/infrastructure/CustomOffsetDateTimeDeserializer.java
@@ -1,21 +1,0 @@
-package org.ecabs.bookings.infrastructure;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-
-import java.io.IOException;
-import java.time.OffsetDateTime;
-import java.time.format.DateTimeFormatter;
-
-//Spring was complaining about deserialization on OffsetDateTime, but luckily Stackoverflow had the answer
-public class CustomOffsetDateTimeDeserializer extends JsonDeserializer<OffsetDateTime> {
-
-    private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSSSSSXXX");
-
-    @Override
-    public OffsetDateTime deserialize(JsonParser parser, DeserializationContext context) throws IOException, JsonProcessingException {
-        return OffsetDateTime.parse(parser.getText(), this.formatter);
-    }
-}
