--- Bookings/src/main/java/org/ecabs/bookings/infrastructure/CustomOffsetDateTimeSerializer.java
@@ -1,20 +1,0 @@
-package org.ecabs.bookings.infrastructure;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-
-import java.io.IOException;
-import java.time.OffsetDateTime;
-import java.time.format.DateTimeFormatter;
-
-//Spring was complaining about serialization on OffsetDateTime, but luckily Stackoverflow had the answer
-public class CustomOffsetDateTimeSerializer extends JsonSerializer<OffsetDateTime> {
-    private DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSSSSSSSSXXX");
-
-    @Override
-    public void serialize(OffsetDateTime value, JsonGenerator gen, SerializerProvider provider) throws IOException, JsonProcessingException {
-        gen.writeString(value.format(this.formatter));
-    }
-}
