--- Bookings/src/main/java/org/ecabs/bookings/infrastructure/db/AuditDAO.java
@@ -1,5 +1,0 @@
-package org.ecabs.bookings.infrastructure.db;
-
-public interface AuditDAO {
-    public void audit(String event);
-}
