--- Bookings/src/main/java/org/ecabs/bookings/infrastructure/db/HibernateConfig.java
@@ -1,83 +1,0 @@
-package org.ecabs.bookings.infrastructure.db;
-
-import org.apache.tomcat.dbcp.dbcp.BasicDataSource;
-import org.hibernate.SessionFactory;
-import org.hibernate.internal.SessionFactoryImpl;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.orm.hibernate5.HibernateTransactionManager;
-import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import javax.sql.DataSource;
-import java.util.Properties;
-
-@Configuration
-@EnableTransactionManagement
-public class HibernateConfig {
-
-    @Value("${database.connection.url}")
-    String url;
-
-    @Value("${database.connection.username}")
-    String username;
-
-    @Value("${database.connection.password}")
-    String password;
-
-    @Value("${database.connection.driver}")
-    String driverClass;
-
-    @Bean
-    AuditDAO audit(){
-        return new AuditDAOImpl();
-    }
-
-    @Bean
-    WaypointDAO waypoint(){
-        return new WaypointDAOImpl();
-    }
-
-    @Bean
-    BookingDAO booking(){
-        return new BookingDAOImpl();
-    }
-
-    @Bean
-    public LocalSessionFactoryBean sessionFactory() {
-        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
-        sessionFactory.setDataSource(dataSource());
-        sessionFactory.setPackagesToScan("org.ecabs.bookings.domain.db","org.ecabs.bookings.domain.messagebroker");
-        sessionFactory.setHibernateProperties(hibernateProperties());
-        return sessionFactory;
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        BasicDataSource dataSource = new BasicDataSource();
-        dataSource.setDriverClassName(driverClass);
-        dataSource.setUrl(url);
-        dataSource.setUsername(username);
-        dataSource.setPassword(password);
-
-        return dataSource;
-    }
-
-    @Bean
-    public PlatformTransactionManager hibernateTransactionManager() {
-        HibernateTransactionManager transactionManager
-                = new HibernateTransactionManager();
-        transactionManager.setSessionFactory(sessionFactory().getObject());
-        return transactionManager;
-    }
-
-    private final Properties hibernateProperties() {
-        Properties hibernateProperties = new Properties();
-        hibernateProperties.setProperty("hibernate.hbm2ddl.auto", "create-drop");
-        hibernateProperties.setProperty("hibernate.dialect", "org.hibernate.dialect.H2Dialect");
-
-        return hibernateProperties;
-    }
-}
