--- Bookings/src/main/resources/application-test.yml
@@ -1,29 +1,0 @@
-ecabs:
-  message:
-    exchange: MessageExchange
-    booking:
-      edit: BookingEditQueue
-      delete: BookingDeleteQueue
-      add: BookingAddQueue
-      exchange: BookingExchange
-    audit: MessageAuditQueue
-
-database:
-  connection:
-    host: localhost
-    port: 3306
-    username: root
-    password: ebamber
-    type: TCP
-  name: bookings
-  table:
-    audit: audits
-    booking: bookings
-    waypoint: waypoints
-
-spring:
-  rabbitmq:
-    host: localhost
-    port: 5672
-    username: guest
-    password: guest
